Rename algorithms: ALL -> INCREMENTAL, NEW -> DELTA
