+++ .github/workflows/ci.yaml
@@ -87,14 +87,14 @@ jobs:
       run: python sample/solver/local_solver.py
 
     # Algorithm
-    - name: Run sample/algorithm/npp_all.py (local file)
-      run: python sample/algorithm/npp_all.py --input="tests/algorithm/numbers_100.txt"
-    - name: Run sample/algorithm/npp_all.py (GCS and output)
-      run: python sample/algorithm/npp_all.py --input="gs://sawatabi-bucket/numbers_100.txt" --output="tests/algorithm/output_all.txt"
-    - name: Run sample/algorithm/npp_new.py (local file)
-      run: python sample/algorithm/npp_new.py --input="tests/algorithm/numbers_100.txt"
-    - name: Run sample/algorithm/npp_new.py (GCS and output)
-      run: python sample/algorithm/npp_new.py --input="gs://sawatabi-bucket/numbers_100.txt" --output="tests/algorithm/output_new.txt"
+    - name: Run sample/algorithm/npp_incremental.py (local file)
+      run: python sample/algorithm/npp_incremental.py --input="tests/algorithm/numbers_100.txt"
+    - name: Run sample/algorithm/npp_incremental.py (GCS and output)
+      run: python sample/algorithm/npp_incremental.py --input="gs://sawatabi-bucket/numbers_100.txt" --output="tests/algorithm/output_incremental.txt"
+    - name: Run sample/algorithm/npp_delta.py (local file)
+      run: python sample/algorithm/npp_delta.py --input="tests/algorithm/numbers_100.txt"
+    - name: Run sample/algorithm/npp_delta.py (GCS and output)
+      run: python sample/algorithm/npp_delta.py --input="gs://sawatabi-bucket/numbers_100.txt" --output="tests/algorithm/output_delta.txt"
     - name: Run sample/algorithm/npp_window.py (local file)
       run: python sample/algorithm/npp_window.py --input="tests/algorithm/numbers_100.txt"
     - name: Run sample/algorithm/npp_window.py (GCS and output)
