+++ sample/algorithm/npp_delta.py
@@ -22,14 +22,14 @@ import npp_functions
 import sawatabi
 
 
-def npp_all(project=None, topic=None, subscription=None, input_path=None, output_path=None):
+def npp_new(project=None, topic=None, subscription=None, input_path=None, output_path=None):
 
     pipeline_args = ["--runner=DirectRunner"]
     # pipeline_args.append("--save_main_session")  # If save_main_session is true, pickle of the session fails on Windows unit tests
     if project is not None:
         pipeline_args.append("--streaming")
 
-    algorithm_options = {"incremental.size": 10, "output.with_timestamp": True, "output.prefix": "<<<\n", "output.suffix": "\n>>>\n"}
+    algorithm_options = {"window.size": 10, "output.with_timestamp": True, "output.prefix": "<<<\n", "output.suffix": "\n>>>\n"}
 
     if topic is not None:
         input_fn = sawatabi.algorithm.IO.read_from_pubsub_as_number(project=project, topic=topic)
@@ -45,7 +45,7 @@ def npp_all(project=None, topic=None, subscription=None, input_path=None, output
         output_fn = sawatabi.algorithm.IO.write_to_stdout()
 
     # Pipeline creation with Sawatabi
-    pipeline = sawatabi.algorithm.All.create_pipeline(
+    pipeline = sawatabi.algorithm.Delta.create_pipeline(
         algorithm_options=algorithm_options,
         input_fn=input_fn,
         map_fn=npp_functions.npp_mapping,
@@ -69,7 +69,7 @@ def main():
     parser.add_argument("--output", dest="output", help="Path (prefix) to the output file or the object to write to.")
     args = parser.parse_args()
 
-    npp_all(args.project, args.topic, args.subscription, args.input, args.output)
+    npp_new(args.project, args.topic, args.subscription, args.input, args.output)
 
 
 if __name__ == "__main__":
