+++ sawatabi/algorithm/__init__.py
@@ -13,9 +13,9 @@
 # limitations under the License.
 
 from sawatabi.algorithm.abstract_algorithm import AbstractAlgorithm
-from sawatabi.algorithm.all import All
+from sawatabi.algorithm.delta import Delta
+from sawatabi.algorithm.incremental import Incremental
 from sawatabi.algorithm.io import IO
-from sawatabi.algorithm.new import New
 from sawatabi.algorithm.window import Window
 
-__all__ = ["AbstractAlgorithm", "IO", "All", "New", "Window"]
+__all__ = ["AbstractAlgorithm", "IO", "Delta", "Incremental", "Window"]
